Devel

+++ packages/ocr/off_line_text_preprocessing/c_src/off_line_text_preprocessing.h
@@ -35,8 +35,8 @@ namespace OCR {
 
         MatrixFloat *add_asc_desc (ImageFloat     *img,
                 MatrixFloat *line_mat,
-                    int v_treshold = 5,
-                    float h_threshold = 22.0f
+                    float v_treshold = 5,
+                    int h_threshold = 22.0f
                     );
         ImageFloat *normalize_size(ImageFloat *source, float ascender_ratio, float descender_ratio,
                 april_utils::vector<Point2D> ascenders, april_utils::vector<Point2D> upper_baseline,
